Moving CommitBatchAssignment to newly introduced assignment package
