+++ cordonbleu-main/src/main/java/com/benromberg/cordonbleu/main/resource/commit/GroupAssignmentResource.java
@@ -5,8 +5,8 @@ import com.benromberg.cordonbleu.data.model.Team;
 import com.benromberg.cordonbleu.data.model.User;
 import com.benromberg.cordonbleu.main.permission.TeamPermissionGuard;
 import com.benromberg.cordonbleu.main.permission.UserWithPermissions;
-import com.benromberg.cordonbleu.service.commit.CommitBatchAssignment;
-import com.benromberg.cordonbleu.service.commit.CommitBatchAssignmentService;
+import com.benromberg.cordonbleu.service.assignment.CommitBatchAssignment;
+import com.benromberg.cordonbleu.service.assignment.CommitBatchAssignmentService;
 import com.benromberg.cordonbleu.service.commit.CommitService;
 import com.benromberg.cordonbleu.service.user.UserService;
 import com.codahale.metrics.annotation.Timed;
