+++ cordonbleu-service/src/main/java/com/benromberg/cordonbleu/service/assignment/CommitBatchAssignmentService.java
@@ -1,4 +1,4 @@
-package com.benromberg.cordonbleu.service.commit;
+package com.benromberg.cordonbleu.service.assignment;
 
 import com.benromberg.cordonbleu.data.model.Commit;
 import com.benromberg.cordonbleu.data.model.User;
