+++ cordonbleu-service/src/test/java/com/benromberg/cordonbleu/service/commit/CommitServiceTest.java
@@ -12,6 +12,7 @@ import com.benromberg.cordonbleu.data.model.CommitApproval;
 import com.benromberg.cordonbleu.data.model.CommitFixture;
 import com.benromberg.cordonbleu.data.model.Team;
 import com.benromberg.cordonbleu.data.model.User;
+import com.benromberg.cordonbleu.service.assignment.CommitBatchAssignment;
 import com.benromberg.cordonbleu.util.ClockService;
 import com.benromberg.cordonbleu.util.SystemTimeRule;
